requirements.txt
selenium
beautifulsoup4
requests
python-telegram-bot==13.15
schedule
